+++ src/DevBetterWeb.Core/Interfaces/IPaymentHandlerPaymentIntent.cs
@@ -1,0 +1,8 @@
+﻿namespace DevBetterWeb.Core.Interfaces
+{
+  public interface IPaymentHandlerPaymentIntent
+  {
+    string CreatePaymentIntent(int orderAmound);
+    string GetClientSecret(string paymentIntentId);
+  }
+}
