+++ src/DevBetterWeb.Core/Interfaces/IPaymentHandlerPrice.cs
@@ -4,10 +4,4 @@
   {
     int GetPriceAmount(string subscriptionPriceId);
   }
-
-  public interface IPaymentHandlerPaymentIntent
-  {
-    string CreatePaymentIntent(int orderAmound);
-    string GetClientSecret(string paymentIntentId);
-  }
 }
